Plans: Move `PlanStatus` and `PlanFeatures` into separate directories
